feat(#450): parse deck attributes
